split Assets <-> SimfileDirectory
